Prepared code for classifier
We will be using SciKit-Learn on a separate machine to analyze the training data.
We will train the data using a decision tree classifier due to the ability to hardcode rather than having to implement a classification system on the Nao. (Attempting to install any form of binary or libraries to the Nao has proved much more difficult than should be)
We will then convert the decision tree to python code and inject it into the new Classify() function.